refactor : 검색필터부분 하위 컴포넌트로 분리
